+++ test/CHANGELOG.md
@@ -1,5 +1,24 @@
 # test
 
+## 1.0.3
+
+### Patch Changes
+
+-   e6196f0: Fix: site directory
+-   42533ab: README.md を記述
+
+## 1.0.3-rc.1
+
+### Patch Changes
+
+-   README.md を記述
+
+## 1.0.3-rc.0
+
+### Patch Changes
+
+-   Fix: site directory
+
 ## 1.0.2
 
 ### Patch Changes